migrate sanitize-node to typescript
